gcc 4.0 warning fixes for Interlocked* functions.

+++ dlls/dinput/joystick_linux.c
@@ -89,7 +89,7 @@ static const IDirectInputDevice8WVtbl JoystickWvt;
 struct JoystickImpl
 {
         const void                     *lpVtbl;
-        DWORD                           ref;
+        LONG                            ref;
         GUID                            guid;
 	char				dev[32];
